Changed pid function 3

+++ benchmarks/barrel_ecal/scripts/emcal_barrel_pions_electrons_analysis.cxx
@@ -70,7 +70,7 @@ void emcal_barrel_pions_electrons_analysis(const char* input_fname = "sim_output
 
   // Returns the pdgID of the particle
   auto getpid = [](std::vector<dd4pod::Geant4ParticleData> const& input) {
-    std::vector<double> result = {input[2].pdgID, input[3].pdgID};
+    std::vector<int> result = {input[2].pdgID, input[3].pdgID};
     return result;
   };
